+++ src/base/frontend/javascript/view/electron-tabs.js
@@ -1,48 +1,48 @@
-// Credit to Brrd, the creator of Electron Tabs: https://github.com/brrd/electron-tabs
-//////////////////////////////////////////////////////////////////////////////////////
-// Import functions from other scripts
-// import { darkmode } from '../webview/theme-dark.js'
-
 // LocalStorage - Grab options set by end-user
-const hostname = localStorage.getItem('customHostname')
-const theme = localStorage.getItem('theme')
+const Instance = localStorage.getItem('Instance')
 
-// setTimeout is used so we can delay the tab execution by 1 second.
-// If we delay by 0.9s or sooner, the following error will occur
-// "Uncaught SyntaxError: Decimals with leading zeros are not allowed in strict mode."
+// Tab Group
 setTimeout(() => {
-    // Select tab-group
-    const tabGroup = document.querySelector("tab-group")
+  // Select tab-group
+  const TabGroup = document.querySelector("tab-group")
 
-    // New Tab - When "+" is clicked
-    tabGroup.setDefaultTab({
-      src: hostname,
-      active: true,
-      webviewAttributes: {
-        preload: "./frontend/javascript/webview/preload.js",
-      },
-      ready: function (tab) {
-        const webview = tab.webview;
-        webview.addEventListener('page-title-updated', () => {
-          const newTitle = webview.getTitle()
-          tab.setTitle(newTitle);
-        })
-    }})
+  // New Tab - When "+" is clicked
+  TabGroup.setDefaultTab({
+    src: Instance,
+    active: true,
+    webviewAttributes: {
+      preload: "./frontend/javascript/webview/preload.js",
+      allowpopups: true,
+    },
+    ready: function (tab) {
+      TabGroup.on("tab-removed", (tab, tabGroup) => {ATWC()})
+      const webview = tab.webview
+      webview.addEventListener('page-title-updated', () => {
+        const newTitle = webview.getTitle()
+        tab.setTitle(newTitle)
+      })
+    }
+  })
 
-    // Default Tab - On Launch
-    const tab = tabGroup.addTab({
-      src: hostname,
-      active: true,
-      webviewAttributes: {
-        preload: "./frontend/javascript/webview/preload.js",
-      },
-      ready: function (tab) {
-        const webview = tab.webview;
-        webview.addEventListener('page-title-updated', () => {
-          const newTitle = webview.getTitle()
-          tab.setTitle(newTitle);
-        })
-    }})
+  // Default Tab - On Launch
+  const tab = TabGroup.addTab({
+    src: Instance,
+    active: true,
+    webviewAttributes: {
+      preload: "./frontend/javascript/webview/preload.js",
+      allowpopups: true,
+    },
+    ready: function (tab) {
+      TabGroup.on("tab-removed", (tab, tabGroup) => {ATWC()})
+      const webview = tab.webview
+      webview.addEventListener('page-title-updated', () => {
+        const newTitle = webview.getTitle()
+        tab.setTitle(newTitle)
+      })
+    }
+  })
+
+  // Apply Active WebView Console to App Console
+  const webview = document.querySelector("body > tab-group").shadowRoot.querySelector("div > div > webview.visible")
+  webview.addEventListener('console-message', (e) => {console.log(e.message)})
 }, 1000)
-
-